+++ nodejs/eks/nodegroup.ts
@@ -904,9 +904,9 @@ export function createManagedNodeGroup(name: string, args: ManagedNodeGroupOptio
         scalingConfig: pulumi.all([
             args.scalingConfig,
         ]).apply(([config]) => {
-            const desiredSize = config && config.desiredSize || 2;
-            const minSize = config && config.minSize || 1;
-            const maxSize = config && config.maxSize || 2;
+            const desiredSize = config?.desiredSize ?? 2;
+            const minSize = config?.minSize ?? 1;
+            const maxSize = config?.maxSize ?? 2;
             return {
                 desiredSize: desiredSize,
                 minSize: minSize,